BattleState addition and bug fixes
Added state machine for handling battle states to lay the groundwork for GUI based implementations. Also major bug fixes.

+++ src/battle/StageRatios.h
@@ -2,21 +2,20 @@
 
 #include <utility>
 
-inline constexpr std::pair<unsigned int, unsigned int> GetStageRatio(size_t stage) noexcept
+inline constexpr std::pair<int, int> GetStageRatio(int stage) noexcept
 {
     if (stage <= 6)
     {
-        return { 2, 8 - static_cast<unsigned int>(stage) };
+        return { 2, 8 - stage };
     }
-    return { static_cast<unsigned int>(stage) - 4, 2 };
+    return { stage - 4, 2 };
 }
 
-// You can add your accuracy ratio formula here too!
-inline constexpr std::pair<unsigned int, unsigned int> GetAccuracyStageRatio(size_t stage) noexcept 
+inline constexpr std::pair<int, int> GetAccuracyStageRatio(int stage) noexcept
 {
     if (stage <= 6)
     {
-        return { 6, 18 - (static_cast<unsigned int>(stage) * 2) };
+        return { 6, 18 - (stage * 2) };
     }
-    return { 6 + ((static_cast<unsigned int>(stage) - 6) * 2), 6 };
+    return { 6 + ((stage - 6) * 2), 6 };
 }
